RASPBERRY-PI/requirements.txt
RPi.GPIO
Adafruit_DHT
gpsd-py3
azure-storage-queue 